Completed Logging Module

+++ src/main/java/com/vivek/backend/Management/service/impl/UserServiceImpl.java
@@ -5,54 +5,65 @@ import com.vivek.backend.Management.dao.UserDao;
 import com.vivek.backend.Management.dto.SignupDto;
 import com.vivek.backend.Management.dto.UserResponseDto;
 import com.vivek.backend.Management.entity.User;
+import com.vivek.backend.Management.exception.InvalidCredentialsException;
+import com.vivek.backend.Management.exception.NoRecentUsersExceptions;
+import com.vivek.backend.Management.exception.UserNotFoundException;
 import com.vivek.backend.Management.repository.UserRepository;
 import com.vivek.backend.Management.service.JWTService;
 import com.vivek.backend.Management.service.UserService;
 import com.vivek.backend.Management.vo.RecentUserVO;
 import com.vivek.backend.Management.vo.UserVO;
+import org.slf4j.LoggerFactory;
+import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
+
+
 
 @Service
 public class UserServiceImpl implements UserService {
 
-
     private final UserRepository userRepository;
+    private final UserDao userDao;
+    private final JWTService jwtService;
+    private final AuthenticationManager authManager;
 
     @Autowired
-    UserServiceImpl(UserRepository userRepository)
-    {
+    public UserServiceImpl(UserRepository userRepository, UserDao userDao,
+                           JWTService jwtService, AuthenticationManager authManager) {
         this.userRepository = userRepository;
+        this.userDao = userDao;
+        this.jwtService = jwtService;
+        this.authManager = authManager;
     }
 
-    // used for custom queries
-    @Autowired
-    private UserDao userDao;
+    private static final Logger logger = LoggerFactory.getLogger(UserServiceImpl.class);
+    // This gives you access to logging methods like info(), debug(), warn(), and error().
 
 
-    @Autowired
-    JWTService jwtService;
-
-    @Autowired
-    AuthenticationManager authManager;
-
-
-
+    @Override
     public User createUser(User user)
     {
+        logger.info("Creating User with email: {}",user.getEmail());
+         User savedUser = userRepository.save(user);
+         logger.debug("User created with ID: {}", savedUser.getUserId());
 
-        return userRepository.save(user);
+         return savedUser;
+
     }
 
 
 
-
+     @Override
     public UserResponseDto registerUser(SignupDto signupDto) {
+
+        logger.info("Registering new user: {}",signupDto.getEmail());
         User user = User.builder()
                 .firstName(signupDto.getFirstName())
                 .lastName(signupDto.getLastName())
@@ -62,6 +73,8 @@ public class UserServiceImpl implements UserService {
 
         User savedUser = userRepository.save(user);
 
+
+        logger.debug("User Registered successfully with ID: {}",savedUser.getUserId());
         return UserResponseDto.builder()
                 .id(savedUser.getUserId())
                 .fullName(savedUser.getFirstName() + " " + savedUser.getLastName())
@@ -72,12 +85,18 @@ public class UserServiceImpl implements UserService {
 
 
 
-
+    @Override
     public UserResponseDto getUserById(Long id)
     {
+        logger.info("Fetching user with ID: {}",id);
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found with ID: " + id));
+                .orElseThrow(() -> {
+                    logger.error("User not found with ID {}",id);
+                    return new UserNotFoundException("User not found with ID: {} " + id);
+                });
 
+
+        logger.debug("User Found: {}",user.getEmail());
         return UserResponseDto.builder()
                 .id(user.getUserId())
                 .fullName(user.getFirstName() + " " + user.getLastName())
@@ -87,11 +106,14 @@ public class UserServiceImpl implements UserService {
         //return userRepository.findById(id).orElse(null);
     }
 
-
+    @Override
     public List<UserResponseDto> getAllUser() {
 
+        logger.info("Fetching all users");
         List<User> users = userRepository.findAll();
 
+
+        logger.debug("Total users fetched : {}",users.size());
         return users.stream()
                 .map(user -> UserResponseDto.builder()
                         .id(user.getUserId())
@@ -108,31 +130,61 @@ public class UserServiceImpl implements UserService {
 
 
 
+    @Override
+    public String deleteUserById(Long id) {
+        logger.info("Deleting user with ID :{}", id);
 
-    public String deleteUserById(Long id)
-    {
+        boolean exists = userRepository.existsById(id);
+
+        if (!exists) {
+            logger.warn("User with ID {} not found", id);
+            throw new UserNotFoundException("User not found with ID: " + id);
+        }
+
         userRepository.deleteById(id);
-        return "Successfully deleted user with id: " + id;
+        logger.info("User Deleted successfully");
 
+        return "Successfully deleted user with id: " + id;
     }
 
 
-
-
+    @Override
     public String verify(String email, String password)
     {
-        Authentication authentication =
-                authManager.authenticate(new UsernamePasswordAuthenticationToken(email, password));
 
-        if(authentication.isAuthenticated())
-            return jwtService.generateToken(email);
+        logger.info("Verifying credentials for email: {}",email);
+        try{
+            Authentication authentication = authManager.authenticate(
+                    new UsernamePasswordAuthenticationToken(email, password));
+
+            if(authentication.isAuthenticated())
+            {
+                logger.info("LOGIN SUCCESS | email={} | timestamp={}", email, LocalDateTime.now());
+                return jwtService.generateToken(email);
+            }
+
+        }catch (Exception e){
+            logger.warn("LOGIN FAILURE | email={} | timestamp={} | reason={}", email, LocalDateTime.now(), e.getMessage());
+            throw new InvalidCredentialsException("Invalid email or password");
+        }
+
 
         return "fail";
     }
 
     @Override
     public List<RecentUserVO> getRecentUsers() {
-        return userDao.getUsersRegisteredInLast7Days();
+
+        logger.info("Start Fetching users, which enrolled in last 7 days");
+        List<RecentUserVO> recentUsers= userDao.getUsersRegisteredInLast7Days();
+
+        if (recentUsers.isEmpty()) {
+            logger.warn("No users registered in the last 7 days");
+            throw new NoRecentUsersExceptions("No users registered in the last 7 days");
+        }
+
+        logger.info("Fetched Users , those enrolled in last 7 days");
+        return recentUsers;
     }
